fix(daengle): 배포를 위한 빌드 에러 수정
