+++ apps/daengle/src/constants/commons/routes.ts
@@ -42,9 +42,9 @@ export const ROUTES = {
   MYPAGE: '/mypage',
   MYPAGE_USER_INFO: '/mypage/user-profile',
   MYPAGE_USER_INFO_EDIT: '/mypage/user-profile/edit',
-  MAYPAGE_PET_PROFILE: '/mypage/pet-profile',
-  MAYPAGE_PET_PROFILE_EDIT: '/mypage/pet-profile/edit',
-  MAYPAGE_PET_PROFILE_CREATE: '/mypage/pet-profile/create',
+  MYPAGE_PET_PROFILE: '/mypage/pet-profile',
+  MYPAGE_PET_PROFILE_EDIT: '/mypage/pet-profile/edit',
+  MYPAGE_PET_PROFILE_CREATE: '/mypage/pet-profile/create',
   MYPAGE_REVIEWS: '/mypage/reviews',
   MYPAGE_FAVORITES: '/mypage/favorites',
   MYPAGE_PAYMENTS: '/mypage/payments',
