+++ apps/daengle/src/interfaces/auth/index.ts
@@ -55,3 +55,18 @@ export interface SignificantTags {
   tagName: string;
   tag: string;
 }
+
+export interface PetProfileCreateFormType {
+  image: File | null;
+  name: string;
+  birth: number;
+  gender: 'MALE' | 'FEMALE';
+  breed: string;
+  isNeutered: boolean;
+  weight: 'SMALL' | 'MEDIUM' | 'LARGE';
+  groomingExperience: boolean;
+  isBite: boolean;
+  dislikeParts: DislikeParts[];
+  significantTags: SignificantTags[];
+  significant: string;
+}
